Don't activate UI to manage tree relations for bookmarks in the "unsorted bookmarks" folder.
We often store very large number of items into the folder, however we'll never open them as a group of bookmarks via middle click on the folder. On the other hand, if there are too many sibling bookmarks, TST can freeze to initialize the list of siblings. So I decided to disable the UI for bookmarks in the "unsorted bookmarks" folder. This decision is related to the issue #398.