Properly handle mouse events when leaving or entering the sidebar #1992
This new implementation for mouseover and mouseout is based on the
relatedTarget property of the mouse event. It can be used to detect
if an element is entered from a child, parent or from outside the
window. The same is true for leaving elements. This means we no longer
have to keep track which element was entered or left last and directly
send events without a timeout.