Fix mouse events when moving over the close buttons
Moving from the close button of one tab directly to the close button of
another tab without going through the tab element itself was not considered
in 91f5d32e78. This commit fixes this by checking if the tab of the related
target of the event is another valid tab.