Generalize mouse move events to cover all possible cases
Previously, the enter and leave events for a tab were derived from the
(bubbling) mouseover and mouseout events by checking a number of different
cases, like when entering the top level tab element, but not from any of
its children or entering the tab directly without going through another
element. However, it turned out that moving the mouse directly between
child elements of two tabs was not covered, so another case was added [1].

Later, another missing case was discovered: moving the mouse from an
empty part of the tab to directly onto a child element of a tab.

Instead of adding another special case, this commit simplifies all
previously mentioned cases into a single, most general check: We emit
a tab enter event, when the tab or any of its children is entered, but
not from any other of the same tab's children. This also includes the
case where we cannot determine where the mouse came from (eg. from outside
the sidebar).

[1]: a0653c03f14ba2e9ad42162723e89dedd29c42c2