Avoid unnecessary shape transition for custom elements.
`delete` causes to change a Shape (SM)/Hidden Class (V8)/Structure (JSC) in JSVM
and they treats an object which we remove a property from by `detele` as
dictionary mode internally.

If we treat the object as dictionary or hashmap by adding/removeing
properties frequently, this JSVM behavior is nice.
But it would not be effectively if we treats the object as normal class.

Practically, to avoid unnecessary this kind of internal behavior and to stabilize object's shape,

1. We should initialize all properties with the same order for each of times.
2. We should fill with empty value (`null`, `undefined`, empty string, `-1`, or others)
   when we would not like to hold some values in a property.