README.md minor corrections, partial major edition
Essentially: Mozilla bug 1568155 became a duplicate of 1417183. 

Whilst making that minor update, I took the opportunity to make: 

1. a handful of corrections to spelling
2. various changes to grammar
3. suggested improvements to overall readability, including conciseness. 

If my edits for areas (2) and (3) are excessive: apologies! Please cherry-pick whatever suits you. 

I should describe this as a _partial_ major edition, because I have not (yet) paid attention to the page as a whole. (It's now time for me to shift focus to my day job …) I'm happy to discuss; to _kick the ball around_, if you like.