Handle heartbeat messages more efficiently.
Boxing the heartbeat message into an array causes the message to proceed down the general batch message processing path. By posting just the heartbeat message on its own, we can avoid the boxing and unboxing and iterating through the array, with all the memory allocation that entails. It also means the heartbeat messages can be responded to more quickly, which is advantageous in a watchdog system.